pyproject.toml
[project]
name = "prismrerankerv1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars>=1.38.1",
    "torch",
    "torchvision",
    "torchaudio",
    "sentence-transformers>=5.2.3",
    "peft>=0.18.1",
    "voyageai>=0.3.7",
    "openpyxl>=3.1.5",
    "tqdm>=4.67.3",
    "python-dotenv>=1.2.1",
    "huggingface-hub>=0.30.0,<1.0",
    "llama-cpp-python>=0.3.16",
    "flash-attn",
    "vllm>=0.8.5.post1",
    "tiktoken",
    "wandb>=0.25.0",
    "pyyaml>=6.0.3",
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "transformers>=4.52.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "llama-cpp-python-cu124"
url = "https://abetlen.github.io/llama-cpp-python/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }
llama-cpp-python = { index = "llama-cpp-python-cu124" }
flash-attn = { url = "https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.7.16/flash_attn-2.7.4+cu124torch2.6-cp312-cp312-linux_x86_64.whl" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils", "process_data", "rank_evaluate", "train"]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
]
